requirements.txt
paho-mqtt
opencv-python-headless
numpy
requests
python-dotenv
